+++ src/docs/devguide/hugo/site/content/release-notes.md
@@ -1,6 +1,6 @@
 This page describes changes to the Money Transfer Retailer Interface implemented across different releases of the interface.
 
-## v2.9.0
+## v2.11.0
 
 Released xx May 2019
 
@@ -18,6 +18,22 @@ Released xx May 2019
     - `confirmPayment`
     - `reversePayment`
     - `orderStatus`
+
+## v2.10.0
+
+Released 17 May 2019
+
+- Updated base API dependency to v3.20.0
+	- Added new `Account.AccountType` value `CASH_PICKUP` for cash pickup scenarios.
+
+## v2.9.0
+
+Released 16 May 2019
+
+- Updated the `MoneyTransferLookupResponse` model to contain the order references that are returned in the `MoneyTransferAuthResponse`, namely:
+    - `orderRedeemRef` - Reference used by the recipient to redeem the order.
+    - `orderRedeemRefAlt` - An alternate reference used by the recipient to redeem the order.
+    - `orderId` - Reference used by the service provider to uniquely identify the money transfer order on their system.
 
 ## v2.8.0